Fix SQL injection vulnerability
We were building our query strings by interpolation, which would
allow SQL injection via the plate barcode or sample_description.
This replaces the query with a parameterized version, which
avoids the need for us to concatenate the strings ourselves.

The change also initializes the concat_frame with an empty
DataFrame, rather than None, to simplify the logic elsewhere.