Add plates endpoint for box-buster
Closes #147

- Add plates endpoint

Note: This may need revisiting, as currently hits mongo once per plate.
Performance wise this is about 4s per box (depending of course on box
size).

We can likely improve this by performing an aggregate query. This will
be much easier once the positive calls are persisted in mongo
(Which is all happening about the same time as this) this will also
avoid any issues cause if the positive criteria should change.